operator up

+++ epc-web-client/src/main/java/com/epc/web/client/controller/operator/OperatorController.java
@@ -1,8 +1,6 @@
 package com.epc.web.client.controller.operator;
 
 import com.epc.common.Result;
-
-import com.epc.common.constants.Const;
 import com.epc.web.client.controller.common.BaseController;
 import com.epc.web.client.controller.operator.handle.*;
 import com.epc.web.client.controller.operator.query.ClientHandleOperatorCellphone;
@@ -10,12 +8,14 @@ import com.epc.web.client.controller.operator.query.ClientHandleOperatorFindAllB
 import com.epc.web.client.controller.operator.query.ClientHandleOperatorId;
 import com.epc.web.client.controller.supplier.handle.ClientRoleDetailInfo;
 import com.epc.web.client.remoteApi.operator.OperatorClient;
-import com.epc.web.facade.loginuser.dto.LoginUser;
 import com.epc.web.facade.operator.handle.*;
 import com.epc.web.facade.operator.query.HandleOperatorCellphone;
 import com.epc.web.facade.operator.query.HandleOperatorFindAllByName;
 import com.epc.web.facade.operator.query.HandleOperatorId;
 import com.epc.web.facade.operator.vo.OperatorBasicInfoVO;
+import com.epc.web.facade.operator.vo.OperatorBasicVO;
+import com.epc.web.facade.operator.vo.TPurchaserBasicInfoVO;
+import com.epc.web.facade.operator.vo.TSupplierBasicInfoVO;
 import com.epc.web.facade.supplier.handle.RoleDetailInfo;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -59,7 +59,7 @@ public class OperatorController extends BaseController {
 
     /**1
      *
-     * version 2:业务修改，被拉取的设置其默认密码为123456，直接从电话号码 与 密码框进行登陆
+     * version 2:业务修改，被拉取的设置其默认密码为epc1688，并状态为1（拉取状态），直接从电话号码 与 密码框进行登陆,进去就设置密码，将状态改成2（完善信息中）
      *
      *  运营商注册,(有人拉的，手机与名字都有,只需要输入电话，姓名就可以登陆)
      *          (有单独的页面登陆，只需要输入姓名，电话就可以进行登陆，进去直接设置密码，然后完善个人信息，然后下次登陆，就查询这个电话下的这条数据的密码状态是否为空，
@@ -78,22 +78,10 @@ public class OperatorController extends BaseController {
      * 完善运营商信息
      */
     @ApiOperation(value = "2:完善运营商信息)",notes = "donghuan")
-    @PostMapping(value = "/insertCompleteOperatorInfo")
+    @PostMapping(value = "public/insertCompleteOperatorInfo")
     public Result<Boolean> insertCompleteOperatorInfo(@RequestBody ClientRoleDetailInfo clientRoleDetailInfo){
         RoleDetailInfo roleDetailInfo=new RoleDetailInfo();
         BeanUtils.copyProperties(clientRoleDetailInfo,roleDetailInfo);
-        //当前 用户id
-        Long userId = getLoginUser().getUserId();
-        // 当前 登陆人的角色类型（运营商1，供应商3）
-        Integer type = getLoginUser().getType();
-        // 当前登陆人用户角色 （法人，管理员，员工）
-        Integer loginRole = getLoginUser().getLoginRole();
-        if(userId==null || type==null || loginRole==null){
-            return Result.error("从token中获取当前登陆用户id与角色类型");
-        }
-        roleDetailInfo.setSystemRole(type);
-        roleDetailInfo.setSupplierId(userId);
-        roleDetailInfo.setLoginRole(loginRole);
         return operatorClient.insertCompleteOperatorInfo(roleDetailInfo);
     }
 
@@ -108,20 +96,57 @@ public class OperatorController extends BaseController {
     @PostMapping(value = "/createOperatorEmployee")
     public Result<Boolean> createOperatorEmployee(@RequestBody ClientHandleOperatorAddEmployee clientHandleOperatorAddEmployee){
         HandleOperatorAddEmployee handleOperatorAddEmployee=new HandleOperatorAddEmployee();
-        getLoginUser().get
-        handleOperatorAddEmployee.setId(Long.valueOf(loginUser.getBossId()));
         BeanUtils.copyProperties(clientHandleOperatorAddEmployee,handleOperatorAddEmployee);
+        Integer type = getLoginUser().getType();
+        Integer loginRole = getLoginUser().getLoginRole();
+        Long bossId = getLoginUser().getBossId();
+        if(type==null || loginRole==null || bossId==null){
+            return Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorAddEmployee.setSystemRole(type);
+        handleOperatorAddEmployee.setLoginRole(loginRole);
+        handleOperatorAddEmployee.setBossId(bossId);
         return operatorClient.createOperatorEmployee(handleOperatorAddEmployee);
     }
 
+//    /**4
+//     * 依据id查询已经登陆的个人信息(如果是法人，管理员，员工)
+//     */
+//    @ApiOperation(value = "4:依据id查询已经登陆的个人信息",notes = "donghuan")
+//    @PostMapping(value = "/findByNameOperator")
+//    public Result<OperatorBasicInfoVO> findByName(@RequestBody ClientHandleOperatorId clientHandleOperatorId){
+//        HandleOperatorId handleOperatorId=new HandleOperatorId();
+//        BeanUtils.copyProperties(clientHandleOperatorId,handleOperatorId);
+//        Long userId = getLoginUser().getUserId();
+//        Long bossId = getLoginUser().getBossId();
+//        Integer type = getLoginUser().getType();
+//        Integer loginRole = getLoginUser().getLoginRole();
+//        if(userId==null || type==null || loginRole==null || bossId==null){
+//            return Result.success("从redis中获取当前登陆用户信息 异常");
+//        }
+//        handleOperatorId.s
+//        handleOperatorId
+//
+//        return operatorClient.findByName(handleOperatorId);
+//    }
     /**4
-     * 依据id查询已经登陆的个人信息
+     * 依据id查询已经登陆的个人信息(如果是法人，管理员，员工)
      */
     @ApiOperation(value = "4:依据id查询已经登陆的个人信息",notes = "donghuan")
     @PostMapping(value = "/findByNameOperator")
-    public Result<OperatorBasicInfoVO> findByName(@RequestBody ClientHandleOperatorId clientHandleOperatorId){
+    public Result<OperatorBasicVO> findByName(){
         HandleOperatorId handleOperatorId=new HandleOperatorId();
-        BeanUtils.copyProperties(clientHandleOperatorId,handleOperatorId);
+        Long userId = getLoginUser().getUserId();
+//        Long bossId = getLoginUser().getBossId();
+        Integer type = getLoginUser().getType();
+        Integer loginRole = getLoginUser().getLoginRole();
+        if(userId==null || type==null || loginRole==null){
+            return Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorId.setLoginId(userId);
+//        handleOperatorId.setBossId(bossId);
+        handleOperatorId.setSystemRole(type);
+        handleOperatorId.setLoginRole(loginRole);
         return operatorClient.findByName(handleOperatorId);
     }
 
@@ -133,6 +158,13 @@ public class OperatorController extends BaseController {
     public Result<Boolean> updateOperatorEmployeeById(@RequestBody ClientHandleOperatorUpdateEmployeeById clientHandleOperatorUpdateEmployeeById){
         HandleOperatorUpdateEmployeeById handleOperatorUpdateEmployeeById=new HandleOperatorUpdateEmployeeById();
         BeanUtils.copyProperties(clientHandleOperatorUpdateEmployeeById,handleOperatorUpdateEmployeeById);
+        Integer type = getLoginUser().getType();
+        Integer loginRole = getLoginUser().getLoginRole();
+        if(type==null || loginRole==null){
+            return Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorUpdateEmployeeById.setSystemRole(type);
+        handleOperatorUpdateEmployeeById.setLoginRole(loginRole);
         return operatorClient.updateOperatorEmployeeById(handleOperatorUpdateEmployeeById);
     }
 
@@ -148,6 +180,13 @@ public class OperatorController extends BaseController {
     public Result<Boolean> findOperatorRecordByCellphone(@RequestBody ClientHandleOperatorCellphone clientHandleOperatorCellphone){
         HandleOperatorCellphone handleOperatorCellphone=new HandleOperatorCellphone();
         BeanUtils.copyProperties(clientHandleOperatorCellphone,handleOperatorCellphone);
+        Integer type = getLoginUser().getType();
+        Integer loginRole = getLoginUser().getLoginRole();
+        if(type==null || loginRole==null){
+            return Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorCellphone.setSystemRole(type);
+        handleOperatorCellphone.setLoginRole(loginRole);
         return operatorClient.findOperatorRecordByCellphone(handleOperatorCellphone);
     }
 
@@ -159,6 +198,13 @@ public class OperatorController extends BaseController {
     public Result<Boolean> deleteOperatorEmployeeByCellphone(@RequestBody ClientHandleOperatorCellphone clientHandleOperatorCellphone){
         HandleOperatorCellphone handleOperatorCellphone=new HandleOperatorCellphone();
         BeanUtils.copyProperties(clientHandleOperatorCellphone,handleOperatorCellphone);
+        Integer type = getLoginUser().getType();
+        Integer loginRole = getLoginUser().getLoginRole();
+        if(type==null || loginRole==null){
+            return Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorCellphone.setLoginRole(loginRole);
+        handleOperatorCellphone.setSystemRole(type);
         return operatorClient.deleteOperatorEmployeeByCellphone(handleOperatorCellphone);
     }
 
@@ -174,6 +220,13 @@ public class OperatorController extends BaseController {
     public Result<Boolean> deleteOperatorEmployeeById(@RequestBody ClientHandleOperatorIdAndIsDeleted clientHandleOperatorIdAndIsDeleted){
         HandleOperatorIdAndIsDeleted handleOperatorIdAndIsDeleted=new HandleOperatorIdAndIsDeleted();
         BeanUtils.copyProperties(clientHandleOperatorIdAndIsDeleted,handleOperatorIdAndIsDeleted);
+        Integer type = getLoginUser().getType();
+        Integer loginRole = getLoginUser().getLoginRole();
+        if(type==null || loginRole==null){
+            return Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorIdAndIsDeleted.setLoginRole(loginRole);
+        handleOperatorIdAndIsDeleted.setSystemRole(type);
         return operatorClient.deleteOperatorEmployeeById(handleOperatorIdAndIsDeleted);
     }
 
@@ -185,6 +238,13 @@ public class OperatorController extends BaseController {
     public Result<Boolean> updateOperatorEmployeeRoleById(@RequestBody ClientHandleOperatorRole clientHandleOperatorRole){
         HandleOperatorRole handleOperatorRole=new HandleOperatorRole();
         BeanUtils.copyProperties(clientHandleOperatorRole,handleOperatorRole);
+        Integer type = getLoginUser().getType();
+        Integer loginRole = getLoginUser().getLoginRole();
+        if(type==null || loginRole==null){
+            return  Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorRole.setSystemRole(type);
+        handleOperatorRole.setLoginRole(loginRole);
         return operatorClient.updateOperatorEmployeeRoleById(handleOperatorRole);
     }
 
@@ -196,6 +256,13 @@ public class OperatorController extends BaseController {
     public Result<Boolean> updateOperatorEmployeeByisDeleted(@RequestBody ClientHandleOperatorIdAndIsForbidden clientHandleOperatorIdAndIsForbidden){
         HandleOperatorIdAndIsForbidden handleOperatorIdAndIsForbidden=new HandleOperatorIdAndIsForbidden();
         BeanUtils.copyProperties(clientHandleOperatorIdAndIsForbidden,handleOperatorIdAndIsForbidden);
+        Integer type = getLoginUser().getType();
+        Integer loginRole = getLoginUser().getLoginRole();
+        if(type==null || loginRole==null){
+            return  Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorIdAndIsForbidden.setSystemRole(type);
+        handleOperatorIdAndIsForbidden.setLoginRole(loginRole);
         return operatorClient.updateOperatorEmployeeByisDeleted(handleOperatorIdAndIsForbidden);
     }
 
@@ -207,6 +274,13 @@ public class OperatorController extends BaseController {
     public Result<Boolean> forgetPassword(@RequestBody ClientHandleOperatorForgetPassword clientHandleOperatorForgetPassword){
         HandleOperatorForgetPassword handleOperatorForgetPassword=new HandleOperatorForgetPassword();
         BeanUtils.copyProperties(clientHandleOperatorForgetPassword,handleOperatorForgetPassword);
+        Integer type = getLoginUser().getType();
+//        Integer loginRole = getLoginUser().getLoginRole();
+        if(type==null){
+            return  Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorForgetPassword.setSystemRole(type);
+//        handleOperatorForgetPassword.setLoginRole(loginRole);
         return operatorClient.forgetPassword(handleOperatorForgetPassword);
     }
 
@@ -218,40 +292,99 @@ public class OperatorController extends BaseController {
     public Result<List<OperatorBasicInfoVO>> queryOperatorEmployeeAll(@RequestBody ClientHandleOperatorFindAllByName clientHandleOperatorFindAllByName){
         HandleOperatorFindAllByName handleOperatorFindAllByName=new HandleOperatorFindAllByName();
         BeanUtils.copyProperties(clientHandleOperatorFindAllByName,handleOperatorFindAllByName);
+//        Long userId = getLoginUser().getUserId();
+        Long bossId = getLoginUser().getBossId();
+        Integer type = getLoginUser().getType();
+        Integer loginRole = getLoginUser().getLoginRole();
+        if(bossId==null || type==null || loginRole==null){
+            return  Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+//        handleOperatorFindAllByName.setId(userId);
+        handleOperatorFindAllByName.setBossId(bossId);
+        handleOperatorFindAllByName.setSystemRole(type);
+        handleOperatorFindAllByName.setLoginRole(loginRole);
         return operatorClient.queryOperatorEmployeeAll(handleOperatorFindAllByName);
     }
 
     /**15
-     *  运营商新增采购人（包括完善信息）
+     *  运营商新增采购人（第2步：完善信息）
      */
-    @ApiOperation(value = "15:运营商新增采购人（包括完善信息）",notes = "donghuan")
+    @ApiOperation(value = "运营商新增采购人（第2步：完善信息）-15",notes = "donghuan")
     @PostMapping(value = "/createPurchaseByOperator")
     public Result<Boolean> createPurchaseByOperator(@RequestBody ClientHandleCreatePurchaserByOperator clientHandleCreatePurchaserByOperator){
         HandleCreatePurchaserByOperator handleCreatePurchaserByOperator=new HandleCreatePurchaserByOperator();
         BeanUtils.copyProperties(clientHandleCreatePurchaserByOperator,handleCreatePurchaserByOperator);
+        Integer type = getLoginUser().getType();
+        if(type==null){
+            return  Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleCreatePurchaserByOperator.setSystemRole(type);
         return operatorClient.createPurchaseByOperator(handleCreatePurchaserByOperator);
     }
 
-    /**16
-     *  运营商新增采购人（不包括完善信息，只填写姓名，电话）
+    /**15.5
+     *
+     *  查看当前登陆人拉的采购人列表list
+     *      参数:传入当前运营商的id,去采购basic表中去查，看有哪几个采购人是自己拉的
      */
-    @ApiOperation(value = "16:运营商新增采购人（不包括完善信息，只填写姓名，电话）",notes = "donghuan")
+    @ApiOperation(value = "查看当前登陆人拉的采购人列表list--15.5",notes = "donghuan")
+    @PostMapping(value = "/lookPurchaserList")
+    public Result<List<TPurchaserBasicInfoVO>> lookPurchaserList(){
+        HandleOperatorLoginInfo handleOperatorLoginInfo=new HandleOperatorLoginInfo();
+        Long userId = getLoginUser().getUserId();
+        Long bossId = getLoginUser().getBossId();
+        Integer type = getLoginUser().getType();
+        if(userId==null || bossId==null || type==null){
+            return Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorLoginInfo.setBossId(bossId.intValue());
+        handleOperatorLoginInfo.setId(userId);
+        handleOperatorLoginInfo.setSystemRole(type);
+        return operatorClient.lookPurchaserList(handleOperatorLoginInfo);
+    }
+
+    /**16
+     *  运营商新增采购人（第1步：不包括完善信息，只填写姓名，电话，及默认密码epc1688）
+     */
+    @ApiOperation(value = "运营商新增采购人（第1步：不包括完善信息，只填写姓名，电话，及默认密码epc1688--16）",notes = "donghuan")
     @PostMapping(value = "/createPurchaseByOperatorSimple")
     public Result<Boolean> createPurchaseByOperatorSimple(@RequestBody ClientHandleOperatorCreateSupplier clientHandleOperatorCreateSupplier){
         HandleOperatorCreateSupplier handleOperatorCreateSupplier=new HandleOperatorCreateSupplier();
         BeanUtils.copyProperties(clientHandleOperatorCreateSupplier,handleOperatorCreateSupplier);
+        Long userId = getLoginUser().getUserId();
+        Long bossId = getLoginUser().getBossId();
+        Integer type = getLoginUser().getType();
+        Integer loginRole = getLoginUser().getLoginRole();
+        if(userId==null || bossId==null || type==null || loginRole==null){
+            return Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorCreateSupplier.setSystemRole(type);
+        handleOperatorCreateSupplier.setLoginRole(loginRole);
+        handleOperatorCreateSupplier.setId(userId);
+        handleOperatorCreateSupplier.setBossId(bossId);
         return operatorClient.createPurchaseByOperatorSimple(handleOperatorCreateSupplier);
     }
 
 
     /**17
-     *  运营商新增供应商  (不包括完善信息,只填 姓名、电话)
+     *  运营商新增供应商  (第1步：不包括完善信息，只填写姓名，电话，及默认密码epc1688)
      */
-    @ApiOperation(value = "17:运营商新增供应商  (不包括完善信息,只填 姓名、电话)",notes = "donghuan")
+    @ApiOperation(value = "17:运营商新增供应商 (第1步：不包括完善信息，只填写姓名，电话，及默认密码epc1688)",notes = "donghuan")
     @PostMapping(value = "/operatorCreateSupplierSimple")
     public Result<Boolean> operatorCreateSupplierSimple(@RequestBody ClientHandleOperatorCreateSupplier clientHandleOperatorCreateSupplier){
         HandleOperatorCreateSupplier handleOperatorCreateSupplier=new HandleOperatorCreateSupplier();
         BeanUtils.copyProperties(clientHandleOperatorCreateSupplier,handleOperatorCreateSupplier);
+        Long userId = getLoginUser().getUserId();
+        Long bossId = getLoginUser().getBossId();
+        Integer type = getLoginUser().getType();
+        Integer loginRole = getLoginUser().getLoginRole();
+        if(userId==null || bossId==null || type==null || loginRole==null){
+            return Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorCreateSupplier.setId(userId);
+        handleOperatorCreateSupplier.setBossId(bossId);
+        handleOperatorCreateSupplier.setSystemRole(type);
+        handleOperatorCreateSupplier.setLoginRole(loginRole);
         return operatorClient.operatorCreateSupplierSimple(handleOperatorCreateSupplier);
     }
 
@@ -264,9 +397,33 @@ public class OperatorController extends BaseController {
     public Result<Boolean> operatorCreateSupplier(@RequestBody ClientHandleCreatePurchaserByOperator clientHandleCreatePurchaserByOperator){
         HandleCreatePurchaserByOperator handleCreatePurchaserByOperator=new HandleCreatePurchaserByOperator();
         BeanUtils.copyProperties(clientHandleCreatePurchaserByOperator,handleCreatePurchaserByOperator);
+
+        Integer type = getLoginUser().getType();
+        if(type==null){
+            return Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleCreatePurchaserByOperator.setSystemRole(type);
         return operatorClient.operatorCreateSupplier(handleCreatePurchaserByOperator);
     }
 
+    /**19
+     *    查看当前登陆者拉的供应商列表
+     */
+    @ApiOperation(value = "查看当前登陆者拉的供应商列表--19",notes = "donghuan")
+    @PostMapping(value = "lookSupplierList")
+    public Result<List<TSupplierBasicInfoVO>> lookSupplierList(){
+        HandleOperatorLoginInfo handleOperatorLoginInfo=new HandleOperatorLoginInfo();
+        Long userId = getLoginUser().getUserId();
+        Long bossId = getLoginUser().getBossId();
+        Integer type = getLoginUser().getType();
+        if(userId==null || bossId==null || type==null){
+            return Result.error("从redis中获取当前登陆用户信息 异常");
+        }
+        handleOperatorLoginInfo.setSystemRole(type);
+        handleOperatorLoginInfo.setId(userId);
+        handleOperatorLoginInfo.setBossId(bossId.intValue());
+        return operatorClient.lookSupplierList(handleOperatorLoginInfo);
+    }
